+++ Items/Vanity/PumpkingCloak.cs
@@ -8,8 +8,8 @@ namespace Aequus.Items.Vanity
     [AutoloadEquip(EquipType.Back, EquipType.Front)]
     public class PumpkingCloak : ModItem
     {
-        public static sbyte FrontID { get; private set; }
-        public static sbyte BackID { get; private set; }
+        public static int FrontID { get; private set; }
+        public static int BackID { get; private set; }
 
         public override void SetStaticDefaults()
         {
